yet another refactoring
