+++ src/mob/mod.rs
@@ -2,24 +2,23 @@ use std::time::Duration;
 
 use bevy::prelude::*;
 use bevy_xpbd_2d::prelude::*;
+use rand::Rng;
 
 use crate::global::*;
-use crate::map::*;
+use crate::map::BOX_SIZE;
 use crate::player::Player;
 use crate::projectile::Projectile;
 
+use self::infected::*;
+use self::person::*;
+
 pub mod infected;
 pub mod mob_spawner;
+pub mod person;
 
-pub const PERSON_COUNT: i32 = 200;
-pub const PERSON_SPEED: f32 = 20.;
-pub const PERSON_SIZE: f32 = 10.;
-pub const INFECTED_COUNT: i32 = 40;
-pub const INFECTED_HP: i32 = 3;
-pub const INFECTION_ATTEMPT_DELAY: u64 = 200;
-
-#[derive(Component)]
-pub struct Person;
+pub const DEFAULT_MOB_SIZE: f32 = 10.;
+pub const DEFAULT_MOB_HP: i32 = 3;
+pub const DEFAULT_MOB_COLOR: Color = Color::GREEN;
 
 #[derive(Component, Debug)]
 pub struct Stats {
@@ -29,36 +28,58 @@ pub struct Stats {
 impl Default for Stats {
     fn default() -> Self {
         Self {
-            hit_points: INFECTED_HP,
+            hit_points: DEFAULT_MOB_HP,
         }
     }
 }
 
 #[derive(Resource)]
-pub struct PersonDirectionTimer {
+pub struct RandomDirectionTimer {
     pub timer: Timer,
 }
 
-#[derive(Component)]
-pub struct InfectTimer {
-    pub timer: Timer,
+#[derive(Bundle)]
+pub struct MobBundle {
+    sprite_bundle: SpriteBundle,
+    stats: Stats,
+    rigid_body: RigidBody,
+    position: Position,
+    collider: Collider,
 }
 
-impl Default for InfectTimer {
+impl Default for MobBundle {
     fn default() -> Self {
+        let square_sprite = Sprite {
+            color: DEFAULT_MOB_COLOR,
+            custom_size: Some(Vec2 {
+                x: DEFAULT_MOB_SIZE,
+                y: DEFAULT_MOB_SIZE,
+            }),
+            ..default()
+        };
+
+        let mut rng = rand::thread_rng();
+        let posx = rng.gen_range(-BOX_SIZE..=BOX_SIZE);
+        let posy = rng.gen_range(-BOX_SIZE..=BOX_SIZE);
+
         Self {
-            timer: Timer::new(
-                Duration::from_millis(INFECTION_ATTEMPT_DELAY),
-                TimerMode::Repeating,
-            ),
+            sprite_bundle: SpriteBundle {
+                sprite: square_sprite,
+                transform: Transform::from_translation(Vec3::new(posx, posy, 0.)),
+                ..default()
+            },
+            stats: Stats { hit_points: 1 },
+            rigid_body: RigidBody::Dynamic,
+            position: Position(Vec2 { x: posx, y: posy }),
+            collider: Collider::cuboid(DEFAULT_MOB_SIZE, DEFAULT_MOB_SIZE),
         }
     }
 }
 
-pub fn update_person_velocity(
+pub fn update_mob_velocity(
     mut velocity_query: Query<&mut LinearVelocity, (Without<Projectile>, Without<Player>)>,
     time: Res<Time>,
-    mut timer_res: ResMut<PersonDirectionTimer>,
+    mut timer_res: ResMut<RandomDirectionTimer>,
 ) {
     timer_res.timer.tick(time.delta());
